Catch Wordpress posting response errors
Received reports of 500 errors when publishing to Wordpress. Found
errors occuring during RPC call to metaWeblog.newPost. According to
stacktrace the "Content-Length" header of the response did not match
the response.body.bytesize. We've chalked this up to a Wordpress issue
and have wrapped the call in an exception handler prompting users
to verify their posts and if necessary repost.

[Finishes #54115841]