Improve two-column CSS
The `two-wide` system causes problems if there are an odd number of columns,
so use floats and `width: 50%` instead. Also changed `box-sizing` so padding
is calculated more predictably.

The 480px stylesheet was only in charge of eliminating the effect of the
`two-pair` wrapping div, it had no other purpose. So it has been removed.