Drop ruby_version to 2.1.1
This seems to actually work on run.pivotal.io as of today. 2.1.2 fails with some
hilarious error.

Travis-ci claims to look at the .ruby-version in your project root if you do not
have it defined in your travis.yml. This seems like it would reduce one additional
place where we define our ruby version (which is currently pretty annoying).